Refactoring tasks services
